LPS-206064 Extract ScopeFinder service from OAuth2AuthorizationExpandoPortalInstanceLifecycleListener to OAuth2AuthorizationExpandoScopeFinder, extract constant from OAuth2AuthorizationExpandoPortalInstanceLifecycleListener to SAPEntryObjectConstants and apply to usages
