+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/java/com/liferay/osb/faro/web/internal/spi/scope/finder/OAuth2AuthorizationExpandoScopeFinder.java
@@ -1,0 +1,46 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.osb.faro.web.internal.spi.scope.finder;
+
+import com.liferay.oauth2.provider.scope.spi.scope.finder.ScopeFinder;
+import com.liferay.osb.faro.web.internal.constants.FaroSAPConstants;
+import com.liferay.petra.function.transform.TransformUtil;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
+
+import java.util.Collection;
+import java.util.List;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Joao Victor Alves
+ */
+@Component(
+	property = {
+		"osgi.jaxrs.name=Liferay.Analytics.Cloud.REST", "sap.scope.finder=true"
+	},
+	service = ScopeFinder.class
+)
+public class OAuth2AuthorizationExpandoScopeFinder implements ScopeFinder {
+
+	@Override
+	public Collection<String> findScopes() {
+		return _scopeAliasesList;
+	}
+
+	@Activate
+	protected void activate() {
+		_scopeAliasesList = TransformUtil.transformToList(
+			FaroSAPConstants.SAP_ENTRY_OBJECT_ARRAYS,
+			sapEntryObject -> StringUtil.replaceFirst(
+				sapEntryObject[0], "OAUTH2_", StringPool.BLANK));
+	}
+
+	private List<String> _scopeAliasesList;
+
+}
